requirements.txt
pandas==2.3.2
numpy==1.24.3
openpyxl==3.1.5
sqlalchemy==2.0.43
python-dateutil==2.8.2
